Reorganized files and separated code
