+++ Modulite/Screens/WidgetConfiguration/Setup/ViewController/Extensions/WidgetSetupViewController+StyleCollectionViewCellDelegate.swift
