+++ Modulite/Screens/WidgetConfiguration/Setup/ViewController/WidgetSetupViewController.swift
@@ -229,135 +229,3 @@ extension WidgetSetupViewController {
         setSetupViewHasAppsSelected(to: true)
     }
 }
-
-// MARK: - SelectedAppCollectionViewCellDelegate
-extension WidgetSetupViewController: SelectedAppCollectionViewCellDelegate {
-    func selectedAppCollectionViewCellDidPressDelete(_ cell: SelectedAppCollectionViewCell) {
-        guard let indexPath = setupView.selectedAppsCollectionView.indexPath(for: cell) else {
-            print("Could not get IndexPath for app cell")
-            return
-        }
-        
-        didMakeChanges = true
-        
-        let app = viewModel.selectedApps[indexPath.row]
-        viewModel.removeSelectedApp(app)
-        delegate?.widgetSetupViewControllerDidDeselectApp(self, app: app)
-        setupView.selectedAppsCollectionView.performBatchUpdates({ [weak self] in
-            self?.setupView.selectedAppsCollectionView.deleteItems(at: [indexPath])
-            
-        }, completion: { _ in
-            UIView.animate(withDuration: 0.3) { [weak self] in
-                self?.setupView.setNeedsLayout()
-                self?.setupView.layoutIfNeeded()
-            }
-        })
-    }
-}
-
-// MARK: - UICollectionViewDataSource
-extension WidgetSetupViewController: UICollectionViewDataSource {
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        2
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        guard section != 0 else { return 0 }
-        
-        switch collectionView {
-        case setupView.stylesCollectionView: return viewModel.widgetStyles.count
-        case setupView.selectedAppsCollectionView: return viewModel.selectedApps.count
-        default: return 0
-        }
-    }
-    
-    func collectionView(
-        _ collectionView: UICollectionView,
-        cellForItemAt indexPath: IndexPath
-    ) -> UICollectionViewCell {
-        
-        switch collectionView {
-        case setupView.stylesCollectionView:
-            guard let cell = collectionView.dequeueReusableCell(
-                withReuseIdentifier: StyleCollectionViewCell.reuseId,
-                for: indexPath
-            ) as? StyleCollectionViewCell else {
-                fatalError("Could not dequeue StyleCollectionViewCell")
-            }
-            
-            let style = viewModel.widgetStyles[indexPath.row]
-            
-            cell.setup(
-                image: style.previewImage,
-                title: style.name,
-                delegate: self,
-                isPurchased: style.isPurchased
-            )
-            
-            cell.hasSelectionBeenMade = viewModel.isStyleSelected()
-            
-            if style == viewModel.selectedStyle {
-                collectionView.selectItem(at: indexPath, animated: true, scrollPosition: [])
-            }
-            
-            return cell
-            
-        case setupView.selectedAppsCollectionView:
-            guard let cell = collectionView.dequeueReusableCell(
-                withReuseIdentifier: SelectedAppCollectionViewCell.reuseId,
-                for: indexPath
-            ) as? SelectedAppCollectionViewCell else {
-                fatalError("Could not dequeue StyleCollectionViewCell")
-            }
-            
-            cell.setup(with: viewModel.selectedApps[indexPath.row].name)
-            cell.delegate = self
-            
-            return cell
-        
-        default: fatalError("Unsupported View Controller.")
-        }
-    }
-    
-    func collectionView(
-        _ collectionView: UICollectionView,
-        viewForSupplementaryElementOfKind kind: String,
-        at indexPath: IndexPath
-    ) -> UICollectionReusableView {
-        
-        guard kind == UICollectionView.elementKindSectionHeader else {
-            return UICollectionReusableView()
-        }
-        
-        guard let header = collectionView.dequeueReusableSupplementaryView(
-            ofKind: kind,
-            withReuseIdentifier: SetupHeaderReusableCell.reuseId,
-            for: indexPath
-        ) as? SetupHeaderReusableCell else {
-            fatalError("Could not dequeue SetupHeader cell.")
-        }
-        
-        if collectionView === setupView.stylesCollectionView {
-            header.setup(title: .localized(for: .widgetSetupViewStyleHeaderTitle))
-            
-        } else {
-            header.setup(title: .localized(for: .widgetSetupViewAppsHeaderTitle))
-        }
-        return header
-    }
-}
-
-// MARK: - UICollectionViewDelegateFlowLayout
-extension WidgetSetupViewController: UICollectionViewDelegateFlowLayout {
-    func collectionView(
-        _ collectionView: UICollectionView,
-        layout collectionViewLayout: UICollectionViewLayout,
-        sizeForItemAt indexPath: IndexPath
-    ) -> CGSize {
-        let text = viewModel.selectedApps[indexPath.row].name
-        let font = UIFont(textStyle: .title3, weight: .semibold)
-        let size = (text as NSString).size(withAttributes: [NSAttributedString.Key.font: font])
-        
-        return CGSize(width: size.width + 45, height: size.height + 24)
-    }
-}
